server/requirements.txt
flask
flask-cors
python-dotenv
psycopg2-binary
supabase
pyjwt
bcrypt
smtplib
email-validator
python-multipart
razorpay
requests
apscheduler
ultralytics
opencv-python-headless
pillow
numpy
openai